pyproject.toml
[project]
name = "proofstack-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Kirill Momotov",email = "unknownman@live.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=6.0,<7.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-storages (>=1.14.6,<2.0.0)",
    "boto3 (>=1.42.16,<2.0.0)",
    "uuid6 (==2025.0.1)",
    "django-reversion (>=6.1.0,<7.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.10,<0.15.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]
